+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.13.2] - 2020-12-27
+### Changed
+- Added libmdns::Service property called `dns_service` to WebThingServer
+- Rust example publishes mDNS-SD (service discovery) correctly
+
 ## [0.13.1] - 2020-09-23
 ### Changed
 - Update author and URLs to indicate new project home.